Remove fretboard trainer trait
