+++ frontend/src/models/mod.rs
@@ -4,6 +4,3 @@ pub mod storage;
 
 // Re-export shared domain models
 pub use shared::models::*;
-
-// Re-export frontend-specific models
-pub use fretboard_trainer::FretboardTrainerTrait;
